project - Copie/requirements.txt
kafka-python-ng==2.2.2
psycopg[binary]==3.3.2
